# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Absensi Karyawan GPS & Barcode API",
        "description": "API Documentation untuk Sistem Absensi Karyawan dengan GPS dan Barcode Scanner",
        "contact": {
            "name": "Admin Absensi",
            "email": "admin@absensi.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/admin/attendances": {
            "get": {
                "tags": [
                    "Admin - Attendances"
                ],
                "summary": "Daftar absensi",
                "description": "Menampilkan daftar semua data absensi karyawan",
                "operationId": "3e3811fda0ba52d8e23527e8949812c6",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Attendance"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/attendances/report": {
            "get": {
                "tags": [
                    "Admin - Attendances"
                ],
                "summary": "Generate laporan absensi",
                "description": "Generate dan download laporan absensi dalam format PDF berdasarkan filter",
                "operationId": "481ac2d268b5fb602e8208efc8372518",
                "parameters": [
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Filter by month (format: Y-m, contoh: 2024-01)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "week",
                        "in": "query",
                        "description": "Filter by week (format: Y-W##, contoh: 2024-W01)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "division",
                        "in": "query",
                        "description": "Filter by division ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "jobTitle",
                        "in": "query",
                        "description": "Filter by job title ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - Returns PDF file",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/admin/employees": {
            "get": {
                "tags": [
                    "Admin - Employees"
                ],
                "summary": "Daftar karyawan",
                "description": "Menampilkan daftar semua karyawan",
                "operationId": "226b1a8c4a86f57827e52d2255b59d00",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/employees/{id}/verify": {
            "post": {
                "tags": [
                    "Admin - Employees"
                ],
                "summary": "Verifikasi akun karyawan",
                "description": "Mengaktifkan dan memverifikasi email karyawan",
                "operationId": "d6a7170e5c278d40329c06a7bf554da9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID (ULID)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect back dengan flash message"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/employees/{id}/deactivate": {
            "post": {
                "tags": [
                    "Admin - Employees"
                ],
                "summary": "Nonaktifkan akun karyawan",
                "description": "Menonaktifkan akun karyawan dengan menghapus email verification",
                "operationId": "41909b0ba6a95bd918a071e53f43ccc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID (ULID)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect back dengan flash message"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/employees/{id}": {
            "delete": {
                "tags": [
                    "Admin - Employees"
                ],
                "summary": "Hapus akun karyawan",
                "description": "Menghapus akun karyawan secara permanen. Tidak dapat menghapus admin atau akun sendiri.",
                "operationId": "42da6948a89b842d5aae59fd97faf984",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID (ULID)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect back dengan flash message"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Cannot delete admin or self"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "description": "Common Swagger Schemas",
                        "type": "string",
                        "example": "01JKNSK1234567890"
                    },
                    "nip": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "group": {
                        "type": "string",
                        "enum": [
                            "user",
                            "admin",
                            "superadmin"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "approved"
                    },
                    "phone": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female"
                        ]
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "birth_place": {
                        "type": "string",
                        "example": "Jakarta"
                    },
                    "address": {
                        "type": "string",
                        "example": "Jl. Example No. 123"
                    },
                    "city": {
                        "type": "string",
                        "example": "Jakarta"
                    },
                    "education_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "division_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "job_title_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "datetime",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            },
            "Attendance": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "string",
                        "example": "01JKNSK1234567890"
                    },
                    "barcode_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-04"
                    },
                    "time_in": {
                        "type": "string",
                        "format": "time",
                        "example": "08:00:00"
                    },
                    "time_out": {
                        "type": "string",
                        "format": "time",
                        "example": "17:00:00",
                        "nullable": true
                    },
                    "shift_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double",
                        "example": -6.2088
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double",
                        "example": 106.8456
                    },
                    "status": {
                        "type": "string",
                        "example": "present"
                    },
                    "note": {
                        "type": "string",
                        "nullable": true
                    },
                    "attachment": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}